+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+* Default editor is no longer hard-coded, but it's taken from `EDITOR` or `VISUAL` env variables. It still can be overwritten using `editor` field in the config file.
+
 ## [0.12.0] -- 2025-05-22
 
 ### Added